refactor rdb related code to `storage/rdb` folder.
